feat: Move search functionality to the AdventureContainer component
